CI: Troubleshoot Appveyor Failure
Add additional code to attempt to find a solution to the Win81 Appveyor
failure -after- unit tests are done executing when tests succeed on
Win10.

Next will be to try and get the context of the object before release, or
not reference count the JavaScriptRuntimeSafeHandle object, but the
error didn't start at that point.