CI: Got to the bottom of the issue
Apparently windows 8 is a bit more finnicky when it comes to accessing
invalid memory. The issue originated in the promise unit tests -- The
way handles were being created caused them to not be part of the
management cycle and when they got disposed, the runtime was already
collected. On Win8 only this caused issues, other platforms were ok,
oddly. This fixes that issue and has more defensive code around the
management of handles.